refactor: promote api.js out of src/
