+++ packages/xsnap/jsconfig.json
@@ -14,5 +14,11 @@
     "strictNullChecks": true,
     "moduleResolution": "node",
   },
-  "include": ["src/**/*.js", "src/**/*.d.ts", "exported.js", "tools/**/*.js"],
+  "include": [
+    "src/**/*.js",
+    "src/**/*.d.ts",
+    "exported.js",
+    "tools/**/*.js",
+    "api.js"
+  ],
 }
